Rename all p.w to p.outln
Usage of p.w was dangerous, since it expects a format string, so for example defining `linkoptions { "/pdbaltpath:%_PDB%" }` would result in a cryptic Lua error, since the "%" sign has a special meaning.
When passing over user input, it's safer to not use a function that feeds into string.format. Please note that similar vulnerability was present even in Apple's iOS, where malformed Wi-Fi network name would crash the device for this exact reason.
Alternative to this would be to replace offending lines with p.w("%s", ......), which I don't think is cleaner.