Add option to generate all paths relative to root.
This adds a new generation mode that keeps all paths relative to the
root script dir.

This forces GCC to write paths in diagnostics relative to the root
script, which allows VS Code to recognize the paths when ctrl-clicking
the paths in compiler diagnostics in terminal.

Without this, we get path relative to the root, `../../foo.c`, which VS
Code cannot recognize and hinders instant code navigation.